I feel like this is outside the scope of this PR but makes app.js prettier and eases my coding around the newslettermodal. created functions file inside utilities
